requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiostream==0.4.5
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.0.0
exceptiongroup==1.1.1
fastapi==0.98.0
frozenlist==1.3.3
grpclib==0.4.3
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.7.0
markdown-it-py==3.0.0
mdurl==0.1.2
modal-client==0.49.2471
multidict==6.0.4
openai==0.27.8
protobuf==4.23.3
pydantic==1.10.9
Pygments==2.15.1
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
sigtools==4.0.1
sniffio==1.3.0
starlette==0.27.0
synchronicity==0.5.3
tblib==2.0.0
tiktoken==0.4.0
toml==0.10.2
tqdm==4.65.0
typer==0.9.0
types-certifi==2021.10.8.3
types-toml==0.10.8.6
typing-extensions==4.6.3
urllib3==2.0.3
watchfiles==0.19.0
yarl==1.9.2
zipp==3.15.0
